pyproject.toml
[project]
name = "get-class-schedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=6.0",
    "ics>=0.7.2",
    "pycryptodome>=3.23.0",
    "pyinstaller>=6.14.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
